requirements.txt
lxml==4.5.0
torchvision==0.5.0
xmljson==0.2.0
scipy==1.4.1
tabulate==0.8.6
numpy==1.18.1
ujson==1.35
matplotlib==3.1.3
imgaug==0.4.0
tqdm==4.43.0
opencv_python==4.2.0.32
efficientnet_pytorch==0.6.3
torch==1.4.0
progressbar33==2.4
PyYAML==5.3.1
scikit-learn==0.21.3
